extract `Mcrain::ClientProvider` module
