+++ lib/mcrain/base.rb
@@ -10,6 +10,7 @@ require 'docker'
 
 module Mcrain
   class Base
+    include ClientProvider
 
     class << self
       attr_writer :server_name
@@ -117,32 +118,6 @@ module Mcrain
       raise NotImplementedError
     end
 
-    def client
-      @client ||= build_client
-    end
-
-    def build_client
-      require client_require
-      yield if block_given?
-      client_class.new(*client_init_args)
-    end
-
-    def client_require
-      raise NotImplementedError
-    end
-
-    def client_class
-      raise NotImplementedError
-    end
-
-    def client_init_args
-      raise NotImplementedError
-    end
-
-    def client_script
-      client
-      "#{client_class.name}.new(*#{client_init_args.inspect})"
-    end
 
     def teardown
       begin
